requirements.txt
Babel==2.12.1
matplotlib==3.8.1
numpy==1.25.0
pandas==2.0.3
seaborn==0.13.2
streamlit_nightly==1.31.2.dev20240302
wordcloud==1.9.3
